moved check terms to create wallet from import/create account
